// v2.client/src/shared/api/generated/models/file-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface FileDto {
    'id'?: string;
    'title': string | null;
    'imageId'?: string | null;
    'content'?: string | null;
    'imageData'?: string | null;
    'locaction'?: string | null;
    'fileType'?: number;
    'gameGenre'?: number | null;
    'subject'?: string | null;
    'year'?: number | null;
    'semester'?: number | null;
    'sourceType'?: number;
    'links'?: string | null;
    'useFileUpload'?: boolean | null;
    'isInstall'?: boolean | null;
}